backend/twitter-monitor/requirements.txt
tweepy==4.14.0
textblob==0.17.1
nltk==3.8.1
pandas==2.0.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
